#01/requirements.txt
pyarrow==12.0.1
uvicorn~=0.21.1
aiosqlite==0.17.0
anyio==3.6.1
atlastk==0.13
cffi==1.15.0
cryptography==37.0.2
fastapi~=0.95.0
greenlet==1.1.2
idna==3.3
iso8601==1.0.2
numpy==1.22.4
pandas==1.4.2
passlib==1.7.4
pycparser==2.21
pydantic==1.9.1
pypika-tortoise==0.1.5
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
tortoise-orm==0.19.1
typing-extensions>=4.5.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0